backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0
python-multipart>=0.0.9
aiofiles>=24.1.0
anthropic>=0.30.0
openai>=1.40.0
google-generativeai>=0.7.0
python-dotenv>=1.0.1
pydantic>=2.8.0
httpx>=0.27.0
tiktoken>=0.7.0
google-api-python-client>=2.147.0
google-auth>=2.36.0
google-auth-httplib2>=0.2.0
httplib2>=0.22.0
